libipt, pttc, ptdump, ptet: mode.exec[if]
Event tracing adds a new IF payload bit to MODE.EXEC giving the status of
EFLAGS.IF.  It also changes the binding to allow FUP in addition to TIP
since IF is changed on STI and CLI instructions that do not generate TIP.

Add support for the new MODE.EXEC payload bit to packet encode, decode, and
dumping.  We do not expose the information as tracing event, yet.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>