libipt, sideband, ptxed: iflags event
Add a new event

    ptev_iflags

for tracing interrupt masking.  The new event needs to be enabled by setting

    enable_iflags_events

in the respective decoder configuration.

Note that Event Tracing needs to be enabled via IA32_RTIT_CTL[31] for iflags
events to be useful.  Otherwise, the trace does not contain all EFLAGS.IF
updates resulting in inconsistent information and may also cause a

    -pte_bad_status_update

error when the decoder compares its internal state against the state
provided in the PSB+ header.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>