libipt, event: split fup event binding
When enqueueing partial events for packets that bind to FUP, we can tell
from looking at the event whether the FUP is consumed by that binding.

It generally is, with TSX being the only exception, where the TSX event may
share the FUP with a subsequent TIP.  In the TSX case, we can tell by
looking at the abort flag.

Future patches will add events that no longer allow that distinction.

Split the FUP event binding into

    evb_fup

for events that bind to the FUP but do not consume it and thus allow it to
bind itself to a subsequent TIP to form an asynchronous branch event, and

    evb_fup_bound

for events that bind to FUP and consume it.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>