script, perf-get-opts: change directory searched for sideband files
In perf-get-opts.bash we search for sideband files in the current working
directory.  This is likely not what we want.

If the -m option is given to provide the name of the primary sideband file
with an absolute or relative directory, use that directry to search for
sideband files.

If the -m option is not given or if it only contains a file name without a
directory, use the directory containing the perf.data file.

This allows the script to be called from anywhere.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>