libipt: fix copy_to_user functions
For compatibility reasons, each decoder features a _to_user() function that
only copies the portion of an event, instruction, block, or packet
understood by the caller based on the object's size passed to the function.

If the object being passed is bigger than the object supported by the
decoder, additional members are zero initialized.

That code forgot to cast the object pointer to a byte pointer and hence used
far too big offsets outside the bounds of the object being passed.  Fix it.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>