pttc: handle short FILENAME_MAX
When scanning a yasm .lst file line, we assume FILENAME_MAX to be at least
1024 and hard-code a sscanf() limit of 1023 characters for the filename
string.

This assumption is wrong on some systems and ICX pointed out a potential
overflow.  Fix that.

We still want to use FILENAME_MAX as the scanned filename may later be used
in file operations, so we chose a slightly more complicated solution that
simply hard-coding the maximum filename size.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>