pttc, test: fix pttc error recognition
Commit

    ad761927 script: remove \r in pttc output to fix test.bash on Windows

extended the pttc command with

    | sed 's/\r\n/\n/g'

to fix line endings on Windows.  This discarded the pttc error code and
instead checked the sed error code.

Use spaces instead of newlines to separate .exp files in the pttc output to
avoid this problem altogether instead of trying to handle it in the test
bash script.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>