pttc: prefer sizeof() over strlen() for string literals
We use strnlen() with a generous maximum length on string literals,
expecting the compiler to evaluate it at compile-time and use the resulting
integer constant.

This may result in

    pttc/src/parse.c:662:9: error: ‘strnlen’ specified bound 1024 exceeds source size 2 [-Werror=stringop-overread]
      662 |         parse_token_aux(input, token, strnlen(token, 1024))
          |         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    pttc/src/parse.c:1049:18: note: in expansion of macro ‘parse_token’
     1049 |         status = parse_token(input, ":");
          |                  ^~~~~~~~~~~

There is not really a need to use strnlen().  Replace with sizeof() and
adjust to account for the size of the termination character.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>