libipt, query: indicate pending event after fetch event failure
The query decoder reads one event ahead.  When that fails, we store the
error code in the query decoder's status to be returned on the next query
decoder call, since we cannot report the error immediately.

We do not indicate a pending event, however, so higher level code is going
to proceed on its own until it reaches the next query location.  This
causes errors with asynchronous events to be reported too late.

Indicate a pending event to have higher level software query for the error
status immediately.  This will be tested in subsequent patches.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>