libipt, block: indicate pending event after fetch event failure
The block decoder reads one event ahead.  When that fails, we store the
error code in the decoder's status to be returned on the next
pt_blk_next() call together with an empty block.

While this should be fine, we end up with different behavior compared to
the insn and query decoders.  Indicate a pending event, instead, to align
decoder behaviors.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>