libipt, evq: peek
Add a new function

    pt_evq_peek()

that works similar to

    pt_evq_find()

except that it does not check the event type and instead returns the first
event of any type matching the event binding.

In other words, this returns the event that

    pt_evq_dequeue()

would have returned, except that it returns a pointer to the original
entry inside the queue rather than a copy, as pt_evq_dequeue() would.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>