libipt, ptdump: ignore unknown EVD packets
The current spec defines three types of EVD packets.  It reserves an
additional 29 types for future use.

EVD packets provide additional information for events, but are not
strictly necessary for decoding the trace, nor for understanding it.  It
is therefore better to ignore unknown EVD packets than to fail decode.

We still diagnose unknown CFE event types.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>