threads: support freebsd in threads fallback
It looks like thread sanitizer does not support C11 threads [1].  To run
with thread sanitizer, one hence needs to build with

    -D__STDC_NO_THREADS__=1

to use the fallback threads implementation in this source tree.

[1]: https://github.com/google/sanitizers/issues/1195#issuecomment-800832134

Alyssa Ross reported in issue #95 that FreeBSD defines
PTHREAD_MUTEX_NORMAL as enum and not as macro, so we'd redefine it as
macro with a value that does not exist on FreeBSD.

Check for the existence of PTHREAD_MUTEX_TIMED_NP as macro before
redefining PTHREAD_MUTEX_NORMAL to it.  This should leave things as they
are on FreeBSD and use the PTHREAD_MUTEX_NORMAL enum.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>