libipt, windows: fix section locking error on file map failure
In case we can open a file, create a file descriptor from the file handle,
but then fail to create a file mapping, we leave the section locked.

Closing the file descriptor also closes the file handle so we exit early
and do not complete the full cleanup sequence.  Just skip closing the file
handle, instead.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>