libipt, block, insn: do not fill in event IPs
When the IP of an event has been suppressed, the block and instruction
decoders try to bind it to an instruction.  If they succeed, they fill in
the IP of that instruction.

This will not work for TRIG packets, which will be added in a subsequent
patch, because it interferes with detecting TRIG anchors.

If the IP of an event was provided in the trace via a FUP packet, that
event provides an anchor IP of a subsequent TRIG packet.  If the IP was
filled in by the decoder, the event does not provide a TRIG anchor IP.

We would need to add an indication to the event to allow higher level
decoders to distinguish those cases.  Or add artificial, high-frequency
TRIG anchor events.

Since events are interleaved with instructions, the event IP is already
obvious from its position in the instruction stream.  Providing it in the
event object was merely for convenience.

Stop providing filled-in event IPs in the event.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>