ptxed: fix potential error mixup
In the inner decoding loop, we use status to indicate errors and turn them
into -pte_internal in case we break out of the decode loop without an
error.

The only exception to this is obtaining the decoder offset, where we use
errcode.  If we ever had an error, there, it would get overwritten by
-pte_internal.

Fix it.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>