libipt: bind pwre to pwrx
The first PWRE in a power session binds to the same IP as a subsequent
EXSTOP.  Any further PWRE is standalone.  And if we only go to C0.1,
there is no EXSTOP, at all, and also the first PWRE is standalone.

Logically, PWRE binds to PWRX, and there can be multiple PWRE if we
descend into deper C-states before waking up again at the PWRX.

To avoid distinguishing the first and subsequent PWRE, we bind it to PWRX.
This also places the first PWRE nicely behind the corresponding EXSTOP.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>