libipt, ptxed: insn on block
The library contains two separate full decoders:

  - the instruction flow decoder
  - the block decoder

The former is layered on top of the query decoder, which itself is layered
on top of the event decoder, while the latter is directly layered on top
of the event decoder.

The drawback of this is that everything needs to be implemented twice in
slightly different ways for the two top-level decoders.

Implement the instruction flow decoder on top of the block decoder.

On the up side, this reduces maintenance effort.

On the down side, this creates some overhead, since instructions are now
length-decoded twice: once when forming a block, and once when iterating
over the instructions in the block.  This is mitigated by the block cache
that avoids decoding instructions in the block decoder.

Since this is not a performant setup in the first place, I expect the
additional overhead to be acceptable.

The only functional change is that pt_insn_next() may return

    -pte_event_ignored

in case an event that hasn't been indicated previously is indicated at an
empty block.  The caller simply needs to call pt_insn_event() to get that
event and proceed normally.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>