ptxed: fix error IP reporting
When the instruction flow decoder fails with an error but is still able to
provide an instruction, ptxed uses the instruction's IP for reporting the
error.

For a similar situation with the block decoder, ptxed tries to determine
the next instruction's IP for reporting the error.

Either approach is more useful for some errors and less useful for others.
Use the error code to determine what IP to report and align the reporting.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>