libipt: resync
Add

    pt_insn_resync()
    pt_blk_resync()

for re-synchronizing onto the trace stream inside the current PSB segment
after a decode error.

The safe way to re-synchronize is via

    pt_insn_sync_forward()
    pt_blk_sync_forward()

at the next PSB to guarantee that the decoder is properly initialized.  In
particular, deferred TIP packets are guaranteed to not span across PSB+,
and IP compression is reset at PSB.  This may, unfortunately, lose a fair
amount of trace.

Instead of searching for the next PSB, pt_*_resync() scans ahead for the
next packet that provides a full or compressed IP.  It applies packets
that affect the decoder state like MODE, PIP, or timing packets, and it
ignores TNT, assuming no deferred TIP.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>