libipt: fix blk and insn timing queries
When implementing the block decoder on top of the event decoder in

    ed9222c7 libipt: blk on evt

we dropped

    -pte_no_time
    -pte_no_cbr

error returns when no TSC/CBR has been seen.

This bug was later inherited by the insn decoder in

    8b90e054 libipt, ptxed: insn on block

when we implemented that on top of the block decoder.

We already track whether we have seen TSC, but we're not using it in
pt_blk_time().  Fix that and add tracking for CBR.

Fixes #116.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>